sales_report/requirements.txt
psycopg2-binary
python-dotenv
fpdf
pandas
ipykernel
altair[all]